requirements.txt
geographiclib==2.0
geopy==2.4.1
mariadb==1.1.11
packaging==24.2
python-dotenv==1.0.1
